fix: added permission for proposal

+++ src/proposals/proposals.controller.ts
@@ -14,6 +14,8 @@ import {
   Req,
   ForbiddenException,
   ConflictException,
+  BadRequestException,
+  InternalServerErrorException,
 } from "@nestjs/common";
 import { Request } from "express";
 import { ProposalsService } from "./proposals.service";
@@ -48,7 +50,6 @@ import {
   IFilters,
   ILimitsFilter,
 } from "src/common/interfaces/common.interface";
-import { AllowAny } from "src/auth/decorators/allow-any.decorator";
 import { plainToInstance } from "class-transformer";
 import { validate, ValidatorOptions } from "class-validator";
 import {
@@ -60,6 +61,7 @@ import {
   proposalsFullQueryExampleFields,
 } from "src/common/utils";
 import { JWTUser } from "src/auth/interfaces/jwt-user.interface";
+import { IDatasetFields } from "src/datasets/interfaces/dataset-filters.interface";
 
 @ApiBearerAuth()
 @ApiTags("proposals")
@@ -72,6 +74,127 @@ export class ProposalsController {
     private caslAbilityFactory: CaslAbilityFactory,
   ) {}
 
+  private generateProposalInstanceForPermissions(
+    proposal: ProposalClass | CreateProposalDto,
+  ): ProposalClass {
+    const proposalInstance = new ProposalClass();
+    proposalInstance.accessGroups = proposal.accessGroups || [];
+    proposalInstance.ownerGroup = proposal.ownerGroup;
+    proposalInstance.proposalId = proposal.proposalId;
+    proposalInstance.email = proposal.email;
+    proposalInstance.isPublished =
+      "isPublished" in proposal ? proposal.isPublished : false;
+
+    return proposalInstance;
+  }
+
+  private async permissionChecker(
+    group: Action,
+    proposal: ProposalClass | CreateProposalDto,
+    request: Request,
+    error: Error,
+  ) {
+    const proposalInstance = this.generateProposalInstanceForPermissions(
+      proposal as ProposalClass,
+    );
+
+    const user: JWTUser = request.user as JWTUser;
+    const ability = this.caslAbilityFactory.createForUser(user);
+
+    try {
+      switch (group) {
+        case Action.ProposalsCreate:
+          return (
+            ability.can(Action.ProposalsCreateAny, ProposalClass) ||
+            ability.can(Action.ProposalsCreateOwner, proposalInstance)
+          );
+        case Action.ProposalsRead:
+          return (
+            ability.can(Action.ProposalsReadAny, ProposalClass) ||
+            ability.can(Action.ProposalsReadOneOwner, proposalInstance) ||
+            ability.can(Action.ProposalsReadOneAccess, proposalInstance) ||
+            ability.can(Action.ProposalsReadOnePublic, proposalInstance)
+          );
+        case Action.ProposalsUpdate:
+          return (
+            ability.can(Action.ProposalsUpdateAny, ProposalClass) ||
+            ability.can(Action.ProposalsUpdateOwner, proposalInstance)
+          );
+        case Action.ProposalsDelete:
+          return (
+            ability.can(Action.ProposalsDeleteAny, ProposalClass) ||
+            ability.can(Action.ProposalsDeleteOwner, proposalInstance)
+          );
+        case Action.ProposalsAttachmentCreate:
+          return (
+            ability.can(Action.ProposalsAttachmentCreateAny, ProposalClass) ||
+            ability.can(Action.ProposalsAttachmentCreateOwner, proposalInstance)
+          );
+        case Action.ProposalsAttachmentRead:
+          return (
+            ability.can(Action.ProposalsAttachmentReadAny, ProposalClass) ||
+            ability.can(
+              Action.ProposalsAttachmentReadOwner,
+              proposalInstance,
+            ) ||
+            ability.can(
+              Action.ProposalsAttachmentReadPublic,
+              proposalInstance,
+            ) ||
+            ability.can(Action.ProposalsAttachmentReadAccess, proposalInstance)
+          );
+        case Action.ProposalsAttachmentUpdate:
+          return (
+            ability.can(Action.ProposalsAttachmentUpdateAny, ProposalClass) ||
+            ability.can(Action.ProposalsAttachmentUpdateOwner, proposalInstance)
+          );
+        case Action.ProposalsAttachmentDelete:
+          return (
+            ability.can(Action.ProposalsAttachmentDeleteAny, ProposalClass) ||
+            ability.can(Action.ProposalsAttachmentDeleteOwner, proposalInstance)
+          );
+
+        default:
+          throw error;
+      }
+    } catch (error) {
+      throw InternalServerErrorException;
+    }
+  }
+
+  private async checkPermissionsForProposal(
+    request: Request,
+    id: string,
+    group: Action,
+  ) {
+    const customError = new ForbiddenException("Unauthorized access");
+
+    const proposal = await this.proposalsService.findOne({
+      proposalId: id,
+    });
+    if (proposal) {
+      await this.permissionChecker(group, proposal, request, customError);
+    }
+
+    return proposal;
+  }
+
+  private async checkPermissionsForProposalCreate(
+    request: Request,
+    proposal: CreateProposalDto,
+    group: Action,
+  ) {
+    const customError = new ForbiddenException(
+      "Unauthorized to create this proposal",
+    );
+    if (!proposal) {
+      throw new BadRequestException("Not able to create this proposal");
+    }
+
+    await this.permissionChecker(group, proposal, request, customError);
+
+    return proposal;
+  }
   updateFiltersForList(
     request: Request,
     mergedFilters: IFilters<ProposalDocument, IProposalFields>,
@@ -99,7 +222,7 @@ export class ProposalsController {
   // POST /proposals
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Create, ProposalClass),
+    ability.can(Action.ProposalsCreate, ProposalClass),
   )
   @UseInterceptors(
     new MultiUTCTimeInterceptor<ProposalClass, MeasurementPeriodClass>(
@@ -127,10 +250,14 @@ export class ProposalsController {
     @Req() request: Request,
     @Body() createProposalDto: CreateProposalDto,
   ): Promise<ProposalClass> {
-    const existingProposal = await this.findById(
+    const proposalDTO = await this.checkPermissionsForProposalCreate(
       request,
-      createProposalDto.proposalId,
+      createProposalDto,
+      Action.ProposalsCreate,
     );
+    const existingProposal = await this.proposalsService.findOne({
+      proposalId: createProposalDto.proposalId,
+    });
 
     if (existingProposal) {
       throw new ConflictException(
@@ -138,10 +265,13 @@ export class ProposalsController {
       );
     }
 
-    return this.proposalsService.create(createProposalDto);
+    return this.proposalsService.create(proposalDTO);
   }
 
-  @AllowAny()
+  @UseGuards(PoliciesGuard)
+  @CheckPolicies((ability: AppAbility) =>
+    ability.can(Action.ProposalsCreate, ProposalClass),
+  )
   @HttpCode(HttpStatus.OK)
   @Post("/isValid")
   @ApiOperation({
@@ -159,7 +289,7 @@ export class ProposalsController {
       "Check if the proposal provided pass validation. It return true if the validation is passed",
   })
   async isValid(
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+    @Req() request: Request,
     @Body() createProposal: unknown,
   ): Promise<{ valid: boolean }> {
     const validatorOptions: ValidatorOptions = {
@@ -179,7 +309,7 @@ export class ProposalsController {
   // GET /proposals
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Read, ProposalClass),
+    ability.can(Action.ProposalsRead, ProposalClass),
   )
   @Get()
   @ApiOperation({
@@ -215,7 +345,7 @@ export class ProposalsController {
   // GET /proposals/fullquery
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Read, ProposalClass),
+    ability.can(Action.ProposalsRead, ProposalClass),
   )
   @Get("/fullquery")
   @ApiOperation({
@@ -276,7 +406,7 @@ export class ProposalsController {
   // GET /proposals/fullfacet
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Read, ProposalClass),
+    ability.can(Action.ProposalsRead, ProposalClass),
   )
   @Get("/fullfacet")
   @ApiOperation({
@@ -329,7 +459,7 @@ export class ProposalsController {
   // GET /proposals/:id
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Read, ProposalClass),
+    ability.can(Action.ProposalsRead, ProposalClass),
   )
   @Get("/:pid")
   @ApiOperation({
@@ -351,33 +481,11 @@ export class ProposalsController {
     @Req() request: Request,
     @Param("pid") proposalId: string,
   ): Promise<ProposalClass | null> {
-    const proposal = await this.proposalsService.findOne({
-      proposalId: proposalId,
-    });
-    const user: JWTUser = request.user as JWTUser;
-
-    if (proposal) {
-      // NOTE: We need ProposalClass instance because casl module can not recognize the type from proposal mongo database model. If other fields are needed can be added later.
-      const proposalInstance = new ProposalClass();
-      proposalInstance._id = proposal._id;
-      proposalInstance.accessGroups = proposal.accessGroups;
-      proposalInstance.ownerGroup = proposal.ownerGroup;
-      proposalInstance.proposalId = proposal.proposalId;
-      proposalInstance.email = proposal.email;
-
-      if (user) {
-        const ability = this.caslAbilityFactory.createForUser(user);
-        const canView =
-          ability.can(Action.Manage, proposalInstance) ||
-          ability.can(Action.Read, proposalInstance);
-
-        if (!canView) {
-          throw new ForbiddenException("Unauthorized access");
-        }
-      } else {
-        throw new ForbiddenException("Unauthorized access");
-      }
-    }
+    const proposal = await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsRead,
+    );
 
     return proposal;
   }
@@ -385,7 +493,7 @@ export class ProposalsController {
   // PATCH /proposals/:pid
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Update, ProposalClass),
+    ability.can(Action.ProposalsUpdate, ProposalClass),
   )
   @UseInterceptors(
     new MultiUTCTimeInterceptor<ProposalClass, MeasurementPeriodClass>(
@@ -415,9 +523,15 @@ export class ProposalsController {
       "Update an existing proposal and return its representation in SciCat",
   })
   async update(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
     @Body() updateProposalDto: UpdateProposalDto,
   ): Promise<ProposalClass | null> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsUpdate,
+    );
     return this.proposalsService.update(
       { proposalId: proposalId },
       updateProposalDto,
@@ -425,9 +539,9 @@ export class ProposalsController {
   }
 
   // DELETE /proposals/:id
-  @UseGuards()
+  @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Delete, ProposalClass),
+    ability.can(Action.ProposalsDelete, ProposalClass),
   )
   @Delete("/:pid")
   @ApiOperation({
@@ -443,14 +557,22 @@ export class ProposalsController {
     status: 200,
     description: "No value is returned",
   })
-  async remove(@Param("pid") proposalId: string): Promise<unknown> {
+  async remove(
+    @Req() request: Request,
+    @Param("pid") proposalId: string,
+  ): Promise<unknown> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsDelete,
+    );
     return this.proposalsService.remove({ proposalId: proposalId });
   }
 
   // POST /proposals/:id/attachments
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Create, Attachment),
+    ability.can(Action.ProposalsAttachmentCreate, Attachment),
   )
   @Post("/:pid/attachments")
   @ApiOperation({
@@ -475,9 +597,15 @@ export class ProposalsController {
       "Create a new attachment for the proposal identified by the pid specified",
   })
   async createAttachment(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
     @Body() createAttachmentDto: CreateAttachmentDto,
   ): Promise<Attachment> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsAttachmentCreate,
+    );
     const createAttachment = {
       ...createAttachmentDto,
       proposalId: proposalId,
@@ -487,7 +615,9 @@ export class ProposalsController {
 
   // GET /proposals/:pid/attachments
   @UseGuards(PoliciesGuard)
-  @CheckPolicies((ability: AppAbility) => ability.can(Action.Read, Attachment))
+  @CheckPolicies((ability: AppAbility) =>
+    ability.can(Action.ProposalsAttachmentRead, Attachment),
+  )
   @Get("/:pid/attachments")
   @ApiOperation({
     summary: "It returns all the attachments for the proposal specified.",
@@ -508,15 +638,21 @@ export class ProposalsController {
       "Array with all the attachments associated with the proposal with the pid specified",
   })
   async findAllAttachments(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
   ): Promise<Attachment[]> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsAttachmentRead,
+    );
     return this.attachmentsService.findAll({ proposalId: proposalId });
   }
 
   // PATCH /proposals/:pid/attachments/:aid
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Update, Attachment),
+    ability.can(Action.ProposalsAttachmentUpdate, Attachment),
   )
   @Patch("/:pid/attachments/:aid")
   @ApiOperation({
@@ -544,10 +680,16 @@ export class ProposalsController {
       "Update values of the attachment with id specified associated with the proposal with the pid specified",
   })
   async findOneAttachmentAndUpdate(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
     @Param("aid") attachmentId: string,
     @Body() updateAttachmentDto: UpdateAttachmentDto,
   ): Promise<Attachment | null> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsAttachmentUpdate,
+    );
     return this.attachmentsService.findOneAndUpdate(
       { _id: attachmentId, proposalId: proposalId },
       updateAttachmentDto,
@@ -557,7 +699,7 @@ export class ProposalsController {
   // DELETE /proposals/:pid/attachments/:aid
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Delete, Attachment),
+    ability.can(Action.ProposalsAttachmentDelete, Attachment),
   )
   @Delete("/:pid/attachments/:aid")
   @ApiOperation({
@@ -583,9 +725,15 @@ export class ProposalsController {
       "Remove the attachment with id specified associated with the proposal with the pid specified",
   })
   async findOneAttachmentAndRemove(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
     @Param("aid") attachmentId: string,
   ): Promise<unknown> {
+    await this.checkPermissionsForProposal(
+      request,
+      proposalId,
+      Action.ProposalsAttachmentDelete,
+    );
     return this.attachmentsService.findOneAndRemove({
       _id: attachmentId,
       proposalId: proposalId,
@@ -595,7 +743,7 @@ export class ProposalsController {
   // GET /proposals/:id/datasets
   @UseGuards(PoliciesGuard)
   @CheckPolicies((ability: AppAbility) =>
-    ability.can(Action.Read, DatasetClass),
+    ability.can(Action.DatasetRead, DatasetClass),
   )
   @Get("/:pid/datasets")
   @ApiOperation({
@@ -618,8 +766,44 @@ export class ProposalsController {
       "Array with all the datasets associated with the proposal with the pid specified",
   })
   async findAllDatasets(
+    @Req() request: Request,
     @Param("pid") proposalId: string,
   ): Promise<DatasetClass[] | null> {
-    return this.datasetsService.findAll({ where: { proposalId } });
+    const user: JWTUser = request.user as JWTUser;
+    const ability = this.caslAbilityFactory.createForUser(user);
+    const canViewAny = ability.can(Action.DatasetReadAny, DatasetClass);
+    const fields: IDatasetFields = JSON.parse("{}");
+
+    if (!canViewAny) {
+      const canViewAccess = ability.can(
+        Action.DatasetReadManyAccess,
+        DatasetClass,
+      );
+      const canViewOwner = ability.can(
+        Action.DatasetReadManyOwner,
+        DatasetClass,
+      );
+      const canViewPublic = ability.can(
+        Action.DatasetReadManyPublic,
+        DatasetClass,
+      );
+      if (canViewAccess) {
+        fields.userGroups = user.currentGroups ?? [];
+        fields.userGroups.push(...user.currentGroups);
+        // fields.sharedWith = user.email;
+      } else if (canViewOwner) {
+        fields.ownerGroup = user.currentGroups ?? [];
+        fields.ownerGroup.push(...user.currentGroups);
+      } else if (canViewPublic) {
+        fields.isPublished = true;
+      }
+    }
+
+    const dataset = await this.datasetsService.fullquery({
+      where: { proposalId },
+      fields: fields,
+    });
+
+    return dataset;
   }
 }
